Set file type to scss, revert #17
This reverts the merged pull request #17.

When the file type is set to 'scss.css', many other Sass-related plugins
will not detect that a SCSS file is indeed a SCSS file. We should
therefor use a file type that's as close to official as it gets, and
since the two official file extensions are .sass and .scss, using 'scss'
as file type is a good choice.